refactor : [005-SIGN-UP-SELLER] 고객, 셀러 패키지 분리
