+++ user-api/src/main/java/com/zerobase/cms/user/application/SignInApplication.java
@@ -5,7 +5,7 @@ import com.zerobase.domain.domain.common.UserType;
 import com.zerobase.cms.user.domain.SignInForm;
 import com.zerobase.cms.user.domain.model.Customer;
 import com.zerobase.cms.user.exception.CustomException;
-import com.zerobase.cms.user.service.CustomerService;
+import com.zerobase.cms.user.service.customer.CustomerService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
