+++ user-api/src/main/java/com/zerobase/cms/user/application/SignUpApplication.java
@@ -5,7 +5,7 @@ import com.zerobase.cms.user.client.mailgun.SendMailForm;
 import com.zerobase.cms.user.domain.SignUpForm;
 import com.zerobase.cms.user.domain.model.Customer;
 import com.zerobase.cms.user.exception.CustomException;
-import com.zerobase.cms.user.service.SignUpCustomerService;
+import com.zerobase.cms.user.service.customer.SignUpCustomerService;
 import lombok.RequiredArgsConstructor;
 import org.apache.commons.lang3.RandomStringUtils;
 import org.springframework.http.ResponseEntity;
