+++ user-api/src/test/java/com/zerobase/cms/user/service/SignUpCustomerServiceTest.java
@@ -2,6 +2,7 @@ package com.zerobase.cms.user.service;
 
 import com.zerobase.cms.user.domain.SignUpForm;
 import com.zerobase.cms.user.domain.model.Customer;
+import com.zerobase.cms.user.service.customer.SignUpCustomerService;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
